airflow/requirements.txt
mlflow 
pandas
scipy
scikit-learn
sqlalchemy
psycopg2-binary
kafka-python